requirements.txt
future==0.18.2
imutils==0.5.4
iso8601==0.1.16
numpy==1.21.2
opencv-python==4.5.3.56
pyserial==3.5
PyYAML==5.4.1
